fix(aait-web-g1): Updated page in home route
